# Apply Soter KDF after PBKDF2 for passphrases