… tests/rust/secure_cell.rs
… tests/themis/themis_secure_cell.c