Instead of using PBKDF2 output directly, pass the resulting *prekey* through an additional round of Soter KDF to associate the encryption key with user-provided associated context as well as the encrypted message length, similar to how symmetric key API does this. Now if different associated context is used for each individual Secure Cell, the encryption key will be different even if PBKDF2 salt somehow ends up the same.

### Compatibility considerations

This changes the encryption algorithm, but it's fine to do this before public release. This is the only chance we have to do this update without it becoming a massive backward-compatibility pain.

The change slightly improves security of the encryption and makes the algorithms a bit more compatible and similar in structure. However, this does not change the API or data format in the slightest.

### Benchmarks

PBKDF2 is the main time sink in passphrase API. It is expected that 1 additional round of Soter KDF (basically, one HMAC-SHA-256 computation) will not have any noticeable effect on the performance of passphrase encryption and decryption, given that we already do 200,000 rounds for PBKDF2.

Here are measurements before and after this change, on my machine with 4 x Intel Core i5-8250U, using Linux with generic OpenSSL 1.1.1g, running only the benchmarks:

| Benchmark | Time, before | Time, after | Change |
| -- | -- | -- | -- |
| Encrypt 64 B | 116.72 ms ± 1.14 ms | 125.26 ms ± 0.66 ms | +8.07% slower |
| Encrypt 4 KB | 118.53 ms ± 0.11 ms | 129.36 ms ± 2.22 ms | +8.58% slower |
| Encrypt 1 MB | 120.44 ms ± 0.65 ms | 130.52 ms ± 0.96 ms | +7.77% slower |
| Decrypt 64 B | 125.63 ms ± 0.30 ms | 128.46 ms ± 0.71 ms | +2.63% slower |
| Decrypt 4 KB | 121.67 ms ± 0.42 ms | 132.60 ms ± 0.41 ms | +9.24% slower |
| Decrypt 1 MB | 122.69 ms ± 0.40 ms | 129.56 ms ± 0.82 ms | +5.84% slower |

<details>
<summary>Benchmark output</summary>

Before:

```
Benchmarking Secure Cell encryption - Seal, passphrase/64: Warming up for 3.0000 s
Secure Cell encryption - Seal, passphrase/64                                                                          
                        time:   [115.50 ms 116.72 ms 117.28 ms]
                        thrpt:  [545.71   B/s 548.30   B/s 554.13   B/s]

Benchmarking Secure Cell encryption - Seal, passphrase/4 KB: Warming up for 3.0000 s
Secure Cell encryption - Seal, passphrase/4 KB                                                                          
                        time:   [118.42 ms 118.53 ms 118.65 ms]
                        thrpt:  [33.713 KiB/s 33.747 KiB/s 33.779 KiB/s]
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) low mild

Benchmarking Secure Cell encryption - Seal, passphrase/1 MB: Warming up for 3.0000 s
Secure Cell encryption - Seal, passphrase/1 MB                                                                          
                        time:   [119.91 ms 120.44 ms 121.20 ms]
                        thrpt:  [8.2505 MiB/s 8.3032 MiB/s 8.3397 MiB/s]

Benchmarking Secure Cell decryption - Seal, passphrase/64: Warming up for 3.0000 s
Secure Cell decryption - Seal, passphrase/64                                                                          
                        time:   [125.29 ms 125.63 ms 125.90 ms]
                        thrpt:  [508.34   B/s 509.42   B/s 510.80   B/s]

Benchmarking Secure Cell decryption - Seal, passphrase/4 KB: Warming up for 3.0000 s
Secure Cell decryption - Seal, passphrase/4 KB                                                                          
                        time:   [121.23 ms 121.67 ms 122.06 ms]
                        thrpt:  [32.771 KiB/s 32.875 KiB/s 32.996 KiB/s]

Benchmarking Secure Cell decryption - Seal, passphrase/1 MB: Warming up for 3.0000 s
Secure Cell decryption - Seal, passphrase/1 MB                                                                          
                        time:   [122.39 ms 122.69 ms 123.18 ms]
                        thrpt:  [8.1182 MiB/s 8.1508 MiB/s 8.1709 MiB/s]
```

After:

```
Benchmarking Secure Cell encryption - Seal, passphrase/64: Warming up for 3.0000 s
Secure Cell encryption - Seal, passphrase/64                                                                          
                        time:   [124.47 ms 125.26 ms 125.79 ms]
                        thrpt:  [508.80   B/s 510.95   B/s 514.20   B/s]
                 change:
                        time:   [+7.0137% +8.0663% +9.1785%] (p = 0.00 < 0.05)
                        thrpt:  [-8.4069% -7.4642% -6.5540%]
                        Performance has regressed.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

Benchmarking Secure Cell encryption - Seal, passphrase/4 KB: Warming up for 3.0000 s
Secure Cell encryption - Seal, passphrase/4 KB                                                                          
                        time:   [127.19 ms 129.36 ms 131.64 ms]
                        thrpt:  [30.385 KiB/s 30.922 KiB/s 31.449 KiB/s]
                 change:
                        time:   [+7.3726% +8.5848% +9.8243%] (p = 0.00 < 0.05)
                        thrpt:  [-8.9454% -7.9061% -6.8664%]
                        Performance has regressed.

Benchmarking Secure Cell encryption - Seal, passphrase/1 MB: Warming up for 3.0000 s
Secure Cell encryption - Seal, passphrase/1 MB                                                                          
                        time:   [129.51 ms 130.52 ms 131.43 ms]
                        thrpt:  [7.6086 MiB/s 7.6614 MiB/s 7.7213 MiB/s]
                 change:
                        time:   [+7.0080% +7.7663% +8.5063%] (p = 0.00 < 0.05)
                        thrpt:  [-7.8394% -7.2066% -6.5491%]
                        Performance has regressed.

Benchmarking Secure Cell decryption - Seal, passphrase/64: Warming up for 3.0000 s
Secure Cell decryption - Seal, passphrase/64                                                                          
                        time:   [127.88 ms 128.46 ms 129.30 ms]
                        thrpt:  [494.97   B/s 498.21   B/s 500.46   B/s]
                 change:
                        time:   [+1.9737% +2.6323% +3.3709%] (p = 0.00 < 0.05)
                        thrpt:  [-3.2610% -2.5648% -1.9355%]
                        Performance has regressed.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

Benchmarking Secure Cell decryption - Seal, passphrase/4 KB: Warming up for 3.0000 s
Secure Cell decryption - Seal, passphrase/4 KB                                                                          
                        time:   [132.27 ms 132.60 ms 133.10 ms]
                        thrpt:  [30.053 KiB/s 30.167 KiB/s 30.241 KiB/s]
                 change:
                        time:   [+8.8970% +9.2366% +9.6182%] (p = 0.00 < 0.05)
                        thrpt:  [-8.7743% -8.4556% -8.1701%]
                        Performance has regressed.
Found 3 outliers among 10 measurements (30.00%)
  1 (10.00%) low mild
  2 (20.00%) high mild

Benchmarking Secure Cell decryption - Seal, passphrase/1 MB: Warming up for 3.0000 s
Secure Cell decryption - Seal, passphrase/1 MB                                                                          
                        time:   [128.97 ms 129.56 ms 130.62 ms]
                        thrpt:  [7.6556 MiB/s 7.7184 MiB/s 7.7540 MiB/s]
                 change:
                        time:   [+5.3473% +5.8421% +6.4588%] (p = 0.00 < 0.05)
                        thrpt:  [-6.0670% -5.5196% -5.0759%]
                        Performance has regressed.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
```

</details>

Well, this is intredesting 🤔 But not very relevant, given the noise brackets.

It doesn't seem like *that* much. I'd expect some variation, but not a consistent performance drop across all input sizes. I guess I suck at microbenchmarks and by environment is noisy. I'll keep looking into it.

Nevertheless, it's great that we *have* benchmarks at all.

## Checklist

- [X] Change is covered by automated tests
- [X] Benchmark results are attached
- [X] The [coding guidelines] are followed
- [X] Public API has proper documentation
- [X] ~~Changelog is updated~~ (nothing notable happens here)
   Original passphrase encryption implemented in #577, this PR cross-reference is enough.

[coding guidelines]: https://github.com/cossacklabs/themis/blob/master/CONTRIBUTING.md