@ilammy 2020-05-15
Alrighty. I realized that I *do* suck at benchmarks: I've been running them on a laptop with CPU frequency scaling enabled. However...

After having the CPUs run at their full 3.4 GHz glory and collecting 50 samples instead of 10 (urgh, 15 minutes for one benchmark run) — I still see consistent 6-8% drop in performance.

<details>
<summary>Benchmark output</summary>

Before:

```
Benchmarking Secure Cell encryption - Seal, passphrase/64: Warming up for 3.0000 s
Secure Cell encryption - Seal, passphrase/64                                                                            
                        time:   [114.10 ms 114.58 ms 114.98 ms]
                        thrpt:  [556.61   B/s 558.56   B/s 560.93   B/s]

Benchmarking Secure Cell encryption - Seal, passphrase/4 KB: Warming up for 3.0000 s
Secure Cell encryption - Seal, passphrase/4 KB                                                                            
                        time:   [116.22 ms 116.46 ms 116.70 ms]
                        thrpt:  [34.276 KiB/s 34.347 KiB/s 34.418 KiB/s]

Benchmarking Secure Cell encryption - Seal, passphrase/1 MB: Warming up for 3.0000 s
Secure Cell encryption - Seal, passphrase/1 MB                                                                            
                        time:   [114.47 ms 114.62 ms 114.82 ms]
                        thrpt:  [8.7095 MiB/s 8.7243 MiB/s 8.7358 MiB/s]
Found 7 outliers among 50 measurements (14.00%)
  5 (10.00%) high mild
  2 (4.00%) high severe
```

```
Benchmarking Secure Cell decryption - Seal, passphrase/64: Warming up for 3.0000 s
Secure Cell decryption - Seal, passphrase/64                                                                            
                        time:   [114.03 ms 114.17 ms 114.31 ms]
                        thrpt:  [559.87   B/s 560.55   B/s 561.25   B/s]
Found 1 outliers among 50 measurements (2.00%)
  1 (2.00%) high mild

Benchmarking Secure Cell decryption - Seal, passphrase/4 KB: Warming up for 3.0000 s
Secure Cell decryption - Seal, passphrase/4 KB                                                                            
                        time:   [114.89 ms 115.12 ms 115.43 ms]
                        thrpt:  [34.654 KiB/s 34.746 KiB/s 34.816 KiB/s]
Found 1 outliers among 50 measurements (2.00%)
  1 (2.00%) high mild

Benchmarking Secure Cell decryption - Seal, passphrase/1 MB: Warming up for 3.0000 s
Secure Cell decryption - Seal, passphrase/1 MB                                                                            
                        time:   [116.84 ms 117.59 ms 118.41 ms]
                        thrpt:  [8.4450 MiB/s 8.5041 MiB/s 8.5590 MiB/s]
Found 1 outliers among 50 measurements (2.00%)
  1 (2.00%) high mild
```

After:

```
Benchmarking Secure Cell encryption - Seal, passphrase/64: Warming up for 3.0000 s
Secure Cell encryption - Seal, passphrase/64                                                                            
                        time:   [119.94 ms 120.38 ms 120.82 ms]
                        thrpt:  [529.72   B/s 531.66   B/s 533.59   B/s]
                 change:
                        time:   [+5.7366% +6.5649% +7.3879%] (p = 0.00 < 0.05)
                        thrpt:  [-6.8796% -6.1605% -5.4253%]
                        Performance has regressed.
Found 5 outliers among 50 measurements (10.00%)
  2 (4.00%) low mild
  1 (2.00%) high mild
  2 (4.00%) high severe

Benchmarking Secure Cell encryption - Seal, passphrase/4 KB: Warming up for 3.0000 s
Secure Cell encryption - Seal, passphrase/4 KB                                                                            
                        time:   [124.16 ms 124.53 ms 124.95 ms]
                        thrpt:  [32.013 KiB/s 32.122 KiB/s 32.218 KiB/s]
                 change:
                        time:   [+7.0678% +7.3965% +7.7285%] (p = 0.00 < 0.05)
                        thrpt:  [-7.1741% -6.8871% -6.6013%]
                        Performance has regressed.

Benchmarking Secure Cell encryption - Seal, passphrase/1 MB: Warming up for 3.0000 s
Secure Cell encryption - Seal, passphrase/1 MB                                                                            
                        time:   [124.96 ms 125.33 ms 125.79 ms]
                        thrpt:  [7.9500 MiB/s 7.9790 MiB/s 8.0028 MiB/s]
                 change:
                        time:   [+8.4013% +8.9421% +9.5508%] (p = 0.00 < 0.05)
                        thrpt:  [-8.7182% -8.2081% -7.7501%]
                        Performance has regressed.
Found 1 outliers among 50 measurements (2.00%)
  1 (2.00%) high mild
```

```
Benchmarking Secure Cell decryption - Seal, passphrase/64: Warming up for 3.0000 s
Secure Cell decryption - Seal, passphrase/64                                                                            
                        time:   [122.87 ms 123.25 ms 123.71 ms]
                        thrpt:  [517.33   B/s 519.27   B/s 520.88   B/s]
                 change:
                        time:   [+8.2849% +8.6548% +9.0025%] (p = 0.00 < 0.05)
                        thrpt:  [-8.2590% -7.9654% -7.6510%]
                        Performance has regressed.
Found 13 outliers among 50 measurements (26.00%)
  7 (14.00%) low mild
  5 (10.00%) high mild
  1 (2.00%) high severe

Benchmarking Secure Cell decryption - Seal, passphrase/4 KB: Warming up for 3.0000 s
Secure Cell decryption - Seal, passphrase/4 KB                                                                            
                        time:   [124.28 ms 124.49 ms 124.75 ms]
                        thrpt:  [32.065 KiB/s 32.130 KiB/s 32.186 KiB/s]
                 change:
                        time:   [+7.8544% +8.1609% +8.4637%] (p = 0.00 < 0.05)
                        thrpt:  [-7.8033% -7.5452% -7.2824%]
                        Performance has regressed.

Benchmarking Secure Cell decryption - Seal, passphrase/1 MB: Warming up for 3.0000 s
Secure Cell decryption - Seal, passphrase/1 MB                                                                            
                        time:   [124.74 ms 125.18 ms 125.67 ms]
                        thrpt:  [7.9577 MiB/s 7.9888 MiB/s 8.0164 MiB/s]
                 change:
                        time:   [+7.1820% +7.7639% +8.3215%] (p = 0.00 < 0.05)
                        thrpt:  [-7.6823% -7.2045% -6.7007%]
                        Performance has regressed.
```

</details>

No serious ideas at the moment:

- Are function calls across the library boundary *that* expensive?
- Does it bump into some invisible wall with CPU caches or something?
- Does `soter_kdf()` do something silly? 
- Do I do something silly, like building without optimizations?

It would be nice if someone could reproduce the experiment.

<details>
<summary>How to benchmark</summary>

```bash
# Prepare repo
git clone https://github.com/cossacklabs/themis.git
cd themis
git remote add ilammy https://github.com/ilammy/themis.git
git fetch ilammy

# Build and install master
git checkout master
make
sudo make install

# Build benchmarks
cd benches/themis
cargo bench --no-run
# Wait for Rust to build stuff...

# Run benchmarks (master)
cargo bench -- --sample-size 10 passphrase
# Wait for benchmarks to complete...

# Build and install this branch
git checkout zrtp-goes-in-every-field
cd ../..
make
sudo make install
cd -

# Run benchmarks again
cargo bench -- --sample-size 10 passphrase
# Wait for benchmarks to complete...
```

</details>

> We assume that it has only one round, right?

Yeah, [it should](https://github.com/cossacklabs/themis/blob/master/src/soter/soter_kdf.c). There are basically `soter_hmac` and `soter_wipe` calls but that's about it.

I haven't done any profiling yet so I have no idea where it spends time.