@vixentael 2020-05-15
Benchmarks on my machine:

| Benchmark | Time, before | Time, after | Change |
| -- | -- | -- | -- |
| Encrypt 64 B | 92.575 ms | 92.510 ms | -0.3886% |
| Encrypt 4 KB | 93.777 ms | 92.403 ms | -1.0970% |
| Encrypt 1 MB | 93.169 ms | 92.837 ms | -0.1334% |
| Decrypt 64 B | 92.339 ms | 92.886 ms | +0.6470% |
| Decrypt 4 KB | 92.444 ms | 93.848 ms | +1.1705% |
| Decrypt 1 MB | 93.392 ms | 95.800 ms | +4.5314% |




<details>
<summary>Benchmark output</summary>

Before

```
➜  themis git:(master) cargo bench -- --sample-size 10 passphrase

Benchmarking Secure Cell encryption - Seal, passphrase/64: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.1s.
Secure Cell encryption - Seal, passphrase/64                                                                          
                        time:   [92.321 ms 92.575 ms 92.952 ms]
                        thrpt:  [688.52   B/s 691.33   B/s 693.23   B/s]
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) high mild
Benchmarking Secure Cell encryption - Seal, passphrase/4 KB: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.2s.
Secure Cell encryption - Seal, passphrase/4 KB                                                                          
                        time:   [93.167 ms 93.777 ms 94.563 ms]
                        thrpt:  [42.300 KiB/s 42.654 KiB/s 42.934 KiB/s]
Benchmarking Secure Cell encryption - Seal, passphrase/1 MB: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.2s.
Secure Cell encryption - Seal, passphrase/1 MB                                                                          
                        time:   [92.730 ms 93.169 ms 93.891 ms]
                        thrpt:  [10.651 MiB/s 10.733 MiB/s 10.784 MiB/s]
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

Benchmarking Secure Cell decryption - Seal, passphrase/64: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.1s.
Secure Cell decryption - Seal, passphrase/64                                                                          
                        time:   [92.035 ms 92.339 ms 92.626 ms]
                        thrpt:  [690.95   B/s 693.10   B/s 695.39   B/s]
Benchmarking Secure Cell decryption - Seal, passphrase/4 KB: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.1s.
Secure Cell decryption - Seal, passphrase/4 KB                                                                          
                        time:   [92.262 ms 92.444 ms 92.564 ms]
                        thrpt:  [43.213 KiB/s 43.269 KiB/s 43.355 KiB/s]
Benchmarking Secure Cell decryption - Seal, passphrase/1 MB: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.2s.
Secure Cell decryption - Seal, passphrase/1 MB                                                                          
                        time:   [92.972 ms 93.392 ms 93.852 ms]
                        thrpt:  [10.655 MiB/s 10.708 MiB/s 10.756 MiB/s]
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
```

After

```
➜  themis git:(zrtp-goes-in-every-field) cargo bench -- --sample-size 10 passphrase

Benchmarking Secure Cell encryption - Seal, passphrase/64: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.1s.
Secure Cell encryption - Seal, passphrase/64                                                                          
                        time:   [92.263 ms 92.510 ms 92.801 ms]
                        thrpt:  [689.64   B/s 691.82   B/s 693.67   B/s]
                 change:
                        time:   [-1.1062% -0.3886% +0.2955%] (p = 0.33 > 0.05)
                        thrpt:  [-0.2946% +0.3901% +1.1186%]
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
Benchmarking Secure Cell encryption - Seal, passphrase/4 KB: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.1s.
Secure Cell encryption - Seal, passphrase/4 KB                                                                          
                        time:   [92.249 ms 92.403 ms 92.644 ms]
                        thrpt:  [43.176 KiB/s 43.289 KiB/s 43.361 KiB/s]
                 change:
                        time:   [-2.1557% -1.0970% +0.0088%] (p = 0.09 > 0.05)
                        thrpt:  [-0.0088% +1.1092% +2.2032%]
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
Benchmarking Secure Cell encryption - Seal, passphrase/1 MB: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.1s.
Secure Cell encryption - Seal, passphrase/1 MB                                                                          
                        time:   [92.638 ms 92.837 ms 93.130 ms]
                        thrpt:  [10.738 MiB/s 10.772 MiB/s 10.795 MiB/s]
                 change:
                        time:   [-0.8029% -0.1334% +0.5553%] (p = 0.72 > 0.05)
                        thrpt:  [-0.5522% +0.1336% +0.8094%]
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

Benchmarking Secure Cell decryption - Seal, passphrase/64: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.1s.
Secure Cell decryption - Seal, passphrase/64                                                                          
                        time:   [92.455 ms 92.886 ms 93.179 ms]
                        thrpt:  [686.85   B/s 689.02   B/s 692.23   B/s]
                 change:
                        time:   [+0.0893% +0.6470% +1.2343%] (p = 0.05 < 0.05)
                        thrpt:  [-1.2193% -0.6428% -0.0893%]
                        Change within noise threshold.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
Benchmarking Secure Cell decryption - Seal, passphrase/4 KB: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.3s.
Secure Cell decryption - Seal, passphrase/4 KB                                                                          
                        time:   [92.957 ms 93.848 ms 94.577 ms]
                        thrpt:  [42.293 KiB/s 42.622 KiB/s 43.031 KiB/s]
                 change:
                        time:   [+0.5580% +1.1705% +1.9286%] (p = 0.00 < 0.05)
                        thrpt:  [-1.8921% -1.1570% -0.5549%]
                        Change within noise threshold.
Benchmarking Secure Cell decryption - Seal, passphrase/1 MB: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 5.2s.
Secure Cell decryption - Seal, passphrase/1 MB                                                                          
                        time:   [95.151 ms 95.800 ms 96.627 ms]
                        thrpt:  [10.349 MiB/s 10.438 MiB/s 10.510 MiB/s]
                 change:
                        time:   [+2.7232% +4.5314% +7.3773%] (p = 0.00 < 0.05)
                        thrpt:  [-6.8704% -4.3350% -2.6510%]
                        Performance has regressed.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
```

</details>